+++ examples/shardingsphere-jdbc-example/mixed-feature-example/sharding-readwrite-splitting-example/sharding-readwrite-splitting-spring-namespace-jpa-example/pom.xml
@@ -26,6 +26,18 @@
         <version>5.0.0-RC1-SNAPSHOT</version>
     </parent>
     <artifactId>sharding-readwrite-splitting-spring-namespace-jpa-example</artifactId>
-    <packaging>pom</packaging>
     <name>${project.artifactId}</name>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shardingsphere.example</groupId>
+            <artifactId>example-spring-jpa</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>shardingsphere-jdbc-core-spring-namespace</artifactId>
+        </dependency>
+    </dependencies>
 </project>
