+++ examples/shardingsphere-jdbc-example/mixed-feature-example/sharding-readwrite-splitting-example/sharding-readwrite-splitting-spring-namespace-mybatis-example/src/main/java/org/apache/shardingsphere/example/sharding/readwrite/splitting/spring/namespace/mybatis/ShardingReadwriteSplittingSpringNamespaceMybatisExample.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.example.sharding.spring.namespace.jpa;
+package org.apache.shardingsphere.example.sharding.readwrite.splitting.spring.namespace.mybatis;
 
 import org.apache.shardingsphere.example.core.api.ExampleExecuteTemplate;
 import org.apache.shardingsphere.example.core.api.service.ExampleService;
@@ -24,12 +24,9 @@ import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 import java.sql.SQLException;
 
-public final class ShardingSpringNamespaceJpaExample {
+public final class ShardingReadwriteSplittingSpringNamespaceMybatisExample {
     
-    private static final String CONFIG_FILE = "META-INF/application-sharding-databases.xml";
-//    private static final String CONFIG_FILE = "META-INF/application-sharding-tables.xml";
-//    private static final String CONFIG_FILE = "META-INF/application-sharding-databases-tables.xml";
-//    private static final String CONFIG_FILE = "META-INF/application-readwrite-splitting.xml";
+    private static final String CONFIG_FILE = "META-INF/application-sharding-readwrite-splitting.xml";
     
     public static void main(final String[] args) throws SQLException {
         try (ConfigurableApplicationContext applicationContext = new ClassPathXmlApplicationContext(CONFIG_FILE)) {
